requirements.txt
aiogram==2.23.1
pandas
python-dotenv